Cleanup and renaming of utilities
